# swagger.yml
# @format

info:
  title: Microfest
  contact:
    name: J. Iqbal
    url: https://github.com/LGUG2Z/microfest
    email: jade@beamery.com
  description: Micro Application Manifest and Configuration Manager
  version: 0.0.1
paths:
  /configuration:
    get:
      summary: Gets the configuration
      operationId: GetConfiguration
      security:
        - APIKeyHeader: []
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/Host"
      responses:
        200:
          $ref: "#/responses/OK"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        500:
          $ref: "#/responses/InternalServerError"
    post:
      summary: Submits a new configuration
      operationId: PostConfiguration
      security:
        - APIKeyHeader: []
      consumes:
        - application/json
      produces:
        - text/plain
      parameters:
        - $ref: "#/parameters/Host"
        - $ref: "#/parameters/Configuration"
      responses:
        201:
          $ref: "#/responses/ConfigurationCreated"
        400:
          $ref: "#/responses/BadRequestBody"
        401:
          $ref: "#/responses/Unauthorized"
        500:
          $ref: "#/responses/InternalServerError"
  /manifest:
    get:
      summary: Gets the manifest
      operationId: GetManifest
      security:
        - APIKeyHeader: []
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/Host"
      responses:
        200:
          $ref: "#/responses/OK"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        500:
          $ref: "#/responses/InternalServerError"
    put:
      summary: Submits a patch to create a new manifest
      operationId: PutManifest
      security:
        - APIKeyHeader: []
      consumes:
        - application/json
      produces:
        - text/plain
      parameters:
        - $ref: "#/parameters/Host"
        - $ref: "#/parameters/Manifest"
      responses:
        201:
          $ref: "#/responses/ManifestCreated"
        400:
          $ref: "#/responses/BadRequestBody"
        404:
          $ref: "#/responses/NotFound"
        401:
          $ref: "#/responses/Unauthorized"
        500:
          $ref: "#/responses/InternalServerError"
    post:
      summary: Submits a new manifest
      operationId: PostManifest
      security:
        - APIKeyHeader: []
      consumes:
        - application/json
      produces:
        - text/plain
      parameters:
        - $ref: "#/parameters/Host"
        - $ref: "#/parameters/Manifest"
      responses:
        201:
          $ref: "#/responses/ManifestCreated"
        400:
          description: Bad Request Body
        401:
          $ref: "#/responses/Unauthorized"
        500:
          $ref: "#/responses/InternalServerError"
schemes:
  - http
swagger: "2.0"
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
definitions:
  principal:
    type: string
parameters:
  Host:
    in: query
    name: host
    type: string
    description: The environment hostname
    required: true
  Manifest:
    in: body
    name: manifest
    description: The manifest to submit
    schema:
      type: object
  Configuration:
    in: body
    name: configuration
    description: The configuration to submit
    schema:
      type: object
responses:
  OK:
    description: OK
    schema:
      description: The JSON artifact
      type: object
    headers:
      Cache-Control:
        description: Cache-Control
        type: string

  ManifestCreated:
    description: Manifest Created
    schema:
      type: string
  ConfigurationCreated:
    description: Configuration Created
    schema:
      type: string
  NotFound:
    description: Not Found
    schema:
      type: string
  BadRequestBody:
    description: Bad Request Body
  Unauthorized:
    description: Unauthorized
    headers:
      WWW-Authenticate:
        description: Authorization information is missing or invalid
        type: string
  InternalServerError:
    description: Internal Server Error
    schema:
      type: string